move app.js back to src directory.
