+++ src/App.js
@@ -1,11 +1,11 @@
 import React from "react";
 import "./App.css";
-import {Navigation} from "./Navigation.js";
-import {ViewSelector} from "./ViewSelector.js";
-import {MetricSelector} from "./MetricSelector.js";
-import {VersionSelector} from "./VersionSelector.js";
-import {ChannelSelector} from "./ChannelSelector.js";
-import GC_MS_nightly_62 from "./../data/GC_MS_nightly_62.json";
+import {Navigation} from "./components/Navigation.js";
+import {ViewSelector} from "./components/ViewSelector.js";
+import {MetricSelector} from "./components/MetricSelector.js";
+import {VersionSelector} from "./components/VersionSelector.js";
+import {ChannelSelector} from "./components/ChannelSelector.js";
+import GC_MS_nightly_62 from "./data/GC_MS_nightly_62.json";
 import fetch from "isomorphic-fetch";
 
 class App extends React.Component {
